[bench] Rename bench2 -> bench
