+++ pkg/bench/generator/config.go
