+++ pkg/bench/generator/constant.go
