+++ pkg/bench/generator/constant_test.go
