+++ pkg/bench/pkg.go
@@ -1,0 +1,7 @@
+package bench
+
+import (
+	"github.com/xephonhq/xephon-k/pkg/util"
+)
+
+var log = util.Logger.NewEntryWithPkg("k.bench")
