+++ pkg/bench/reporter/discard.go
