--- pkg/bench2/pkg.go
@@ -1,7 +1,0 @@
-package bench2
-
-import (
-	"github.com/xephonhq/xephon-k/pkg/util"
-)
-
-var log = util.Logger.NewEntryWithPkg("k.bench2")
